// apps/web/src/contracts/generated/sheet/SyncAction.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SyncCopy } from "./SyncCopy";
import type { SyncId } from "./SyncId";
import type { SyncNotice } from "./SyncNotice";
import type { SyncTarget } from "./SyncTarget";
import type { SyncUpload } from "./SyncUpload";

export type SyncAction = { "type": "download" } & SyncTarget | { "type": "upload" } & SyncUpload | { "type": "create" } & SyncId | { "type": "uploadCopy" } & SyncCopy | { "type": "removeLocal" } & SyncId | { "type": "keepLocal" } & SyncNotice | { "type": "deleteRemote" } & SyncTarget | { "type": "keepRemote" } & SyncNotice;